Delete unused fetch calls and organize apiCalls into directories by database endpoints
