+++ screens/Results/Results.js
@@ -4,10 +4,8 @@ import { Body, Button, Card, CardItem } from 'native-base';
 import { Header } from '../../components/Header';
 import { specifyTargetCondition } from '../../utils/helpers/helpers';
 import { ScrollView } from 'react-native-gesture-handler';
-import {
-  getExplanation,
-  getConditionById
-} from '../../utils/apiCalls/apiCalls';
+import { getExplanation } from '../../utils/apiCalls/Infermedica/getExplanation';
+import { getConditionById } from '../../utils/apiCalls/Conditions/getConditionById';
 
 const height = Dimensions.get('window').height;
