+++ utils/apiCalls/DistanceToDoctor/getDistanceToDoctor.test.js
@@ -1,0 +1,3 @@
+import { MAPQUEST_KEY } from 'react-native-dotenv';
+
+const mapquestDirectionsUrl = 'http://www.mapquestapi.com/directions/v2/route';
