+++ utils/apiCalls/DoctorsAndProviders/getAllProviders.test.js
@@ -1,0 +1,3 @@
+import { getAllProviders } from './getAllProviders';
+
+const doctorsUrl = 'https://memd-doc-search.herokuapp.com/api/v1';
