+++ utils/apiCalls/SendEmail/sendEmailReport.test.js
@@ -1,0 +1,1 @@
+import { sendEmailReport } from './sendEmailReport';
